Move `GraphResolveDepths` to its own module
