+++ packages/graph/hash_graph/lib/graph/src/shared/subgraph/depths.rs
@@ -1,0 +1,34 @@
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+use crate::{knowledge::KnowledgeGraphQueryDepth, ontology::OntologyQueryDepth};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct GraphResolveDepths {
+    #[schema(value_type = number)]
+    pub data_type_resolve_depth: OntologyQueryDepth,
+    #[schema(value_type = number)]
+    pub property_type_resolve_depth: OntologyQueryDepth,
+    #[schema(value_type = number)]
+    pub entity_type_resolve_depth: OntologyQueryDepth,
+    #[schema(value_type = number)]
+    // TODO: is this name accurate/satisfactory with the changes we've made?
+    pub link_resolve_depth: KnowledgeGraphQueryDepth,
+    // TODO: what is this?
+    #[schema(value_type = number)]
+    pub link_target_entity_resolve_depth: KnowledgeGraphQueryDepth,
+}
+
+impl GraphResolveDepths {
+    #[must_use]
+    pub const fn zeroed() -> Self {
+        Self {
+            data_type_resolve_depth: 0,
+            property_type_resolve_depth: 0,
+            entity_type_resolve_depth: 0,
+            link_resolve_depth: 0,
+            link_target_entity_resolve_depth: 0,
+        }
+    }
+}
